5066: Fix broken modal
In a different branch, loan type label appears to have been
removed. Result was a broken form that depended on the label.

Now that label names have been removed, we will need to locate
the place where the loan type option set option ids are matched
up to a human understandable key to identify which ids match
which specific loan types.