9159: Do not show the first image in the other images section
If only one image exists and the image is not set as featured, do
not display the image in the other images section. Image will
display as a featured image even though it is not set as the
featured image.

If a featured image is set, the first image, if not the featured
image, will be displayed in the other images section.